Updated the painter class to use size_t.

+++ Source/Painter.cpp
@@ -56,7 +56,7 @@ void Painter::PaintCircle(const Circle& circle, const bool& filled) const {
 	Vector2D cornerPos = Camera::GetCurrentInstance()->GetPosition() + Camera::GetCurrentInstance()->GetExtents().FlipH();
 
 	Vector2D* corners = new Vector2D[n];
-	for (int i=0; i<n; i++) {
+	for (size_t i=0; i<n; i++) {
 		corners[i] = circle.GetCenter() + Vector2D(rA*cos(2*i*M_PI/n), rA*sin(2*i*M_PI/n));
 		corners[i] = (corners[i]-cornerPos).FlipV()*Vector2D::UnitPixelEquivalent;
 	}
@@ -65,7 +65,7 @@ void Painter::PaintCircle(const Circle& circle, const bool& filled) const {
 		SDL_Vertex* vertices = new SDL_Vertex[n]();
 		int* indices = new int[3*n]();
 
-		for (int i=0; i<n; i++) {
+		for (size_t i=0; i<n; i++) {
 			vertices[i].position = {corners[i].x, corners[i].y};
 			vertices[i].color = {this->paintColor->r, this->paintColor->g, this->paintColor->b, this->paintColor->a};
 
@@ -84,7 +84,7 @@ void Painter::PaintCircle(const Circle& circle, const bool& filled) const {
 		GameManager::GetCurrentManager()->ScheduleColorReset();
 		SDL_SetRenderDrawColor(boundedRenderer, this->paintColor->r, this->paintColor->g, this->paintColor->b, this->paintColor->a);
 
-		for (int i=0; i<n; i++) {
+		for (size_t i=0; i<n; i++) {
 			SDL_RenderDrawLine(boundedRenderer, corners[i].x, corners[i].y, corners[(i+1)%n].x, corners[(i+1)%n].y);
 		}
 	}
@@ -105,13 +105,13 @@ void Painter::PaintRectangle(const Rectangle& rectangle, const bool& filled) con
 	Vector2D cornerPos = Camera::GetCurrentInstance()->GetPosition() + Camera::GetCurrentInstance()->GetExtents().FlipH();
 	Vector2D corners[] = {center+tExt, center+tExtF, center-tExt, center-tExtF};
 
-	for (int i=0; i<4; i++) corners[i] = (corners[i]-cornerPos).FlipV()*Vector2D::UnitPixelEquivalent;
+	for (size_t i=0; i<4; i++) corners[i] = (corners[i]-cornerPos).FlipV()*Vector2D::UnitPixelEquivalent;
 
 	if (filled) {
 		SDL_Vertex vertices[4];
 		int indices[6];
 
-		for (int i=0; i<4; i++) {
+		for (size_t i=0; i<4; i++) {
 			vertices[i].position = {corners[i].x, corners[i].y};
 			vertices[i].color = {this->paintColor->r, this->paintColor->g, this->paintColor->b, this->paintColor->a};
 
@@ -128,15 +128,15 @@ void Painter::PaintRectangle(const Rectangle& rectangle, const bool& filled) con
 		GameManager::GetCurrentManager()->ScheduleColorReset();
 		SDL_SetRenderDrawColor(boundedRenderer, this->paintColor->r, this->paintColor->g, this->paintColor->b, this->paintColor->a);
 
-		for (int i=0; i<4; i++) {
+		for (size_t i=0; i<4; i++) {
 			SDL_RenderDrawLine(boundedRenderer, corners[i].x, corners[i].y, corners[(i+1)%4].x, corners[(i+1)%4].y);
 		}
 	}
 }
 
 void Painter::PaintJointShape(const JointShape& jointShape, const bool& filled) const {
-	for (int i = 0; i < jointShape.GetCircleCount(); i++) this->PaintCircle(jointShape.GetCircle(i));
-	for (int i = 0; i < jointShape.GetRectangleCount(); i++) this->PaintRectangle(jointShape.GetRectangle(i));
+	for (size_t i = 0; i < jointShape.GetCircleCount(); i++) this->PaintCircle(jointShape.GetCircle(i));
+	for (size_t i = 0; i < jointShape.GetRectangleCount(); i++) this->PaintRectangle(jointShape.GetRectangle(i));
 }
